Move out vg sort
